Подцепил гугл карты и сделал простой вариант карусели апартаментов

+++ app/src/main/java/ru/iandreyshev/realestate/domain/Rating.kt
@@ -1,6 +1,6 @@
 package ru.iandreyshev.realestate.domain
 
 data class Rating(
-    val average: Float,
+    val average: Double,
     val ratesCount: Int
 )
